separated the constants
