+++ config/multer.js
@@ -1,0 +1,7 @@
+import multer from 'multer';
+
+const upload = multer({ 
+  dest: '/tmp/'
+});
+
+export default upload;
