+++ constants.js
@@ -1,0 +1,6 @@
+export const APPROVAL_DECISIONS = {
+  APPROVE: "approve",
+  REJECT: "reject"
+};
+
+export default APPROVAL_DECISIONS;
