+++ routes/uploadRoutes.js
@@ -1,12 +1,9 @@
 import express from 'express';
-import multer from 'multer';
+import upload from '../config/multer.js';
 import { generateSummary, humanApproval } from '../controllers/summaryController.js';
 
 const router = express.Router();
 
-const upload = multer({ 
-  dest: '/tmp/'
-});
 router.post('/load-document', upload.single('document'), generateSummary);
 router.post('/save-summary', humanApproval);
