+++ utils/loader.js
@@ -1,4 +1,3 @@
-import fs from "fs";
 import path from "path";
 import { PDFLoader } from "@langchain/community/document_loaders/fs/pdf";
 import { TextLoader } from "langchain/document_loaders/fs/text";
@@ -10,15 +9,11 @@ export async function loadFile(filePath) {
     const loader = new TextLoader(filePath.path);
     return await loader.load();
   }
-
   if (ext === ".pdf") {
     const loader = new PDFLoader(filePath.path,{
         splitPages: false,
       });
     return await loader.load();
-
-    
   }
-
   throw new Error("Unsupported file type. Only .txt and .pdf are supported.");
 }
